% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztils.R
\name{glm_pseudor2}
\alias{glm_pseudor2}
\title{glm_pseudoR2}
\usage{
glm_pseudor2(mod)
}
\arguments{
\item{mod}{The model for which to calculate the pseudo R^2}
}
\value{
The pseudo R^2 value of the model
}
\description{
A function for calculating the pseudo R^2 of a glm object
}
\examples{
gmod <- glm(Sepal.Length ~ Petal.Length + Species, data = iris)
glm_pseudor2(gmod)
}
