% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztils.R
\name{multicdf_plot}
\alias{multicdf_plot}
\title{multicdf_plot}
\usage{
multicdf_plot(
  var,
  seq_length = 50,
  distributions = "all",
  palette = "oslo",
  var_name = NULL
)
}
\arguments{
\item{var}{The variable to for which to plot CDFs}

\item{seq_length}{The number of points over which to fit x}

\item{distributions}{The distributions to fit x against}

\item{palette}{The color palette to use on the graph}

\item{var_name}{The variable name to use for x}
}
\value{
A plot showing the CDF of the selected variable against the
selected distributions over the selected sequence length
}
\description{
This function is a version of multicdf_plot from my MultiFitR package
that is customized to use formatting that I like. However, this formatting
uses several packages that I didn't want to include in the
base MultiFitR or MultiFitRgg packages. (scico)
}
