% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztils.R
\name{multiks_cont}
\alias{multiks_cont}
\title{Multiple Kolmogorov-Smirnov Tests for Continuous Variables}
\usage{
multiks_cont(var, distributions = "all")
}
\arguments{
\item{var}{The variable to perform ks tests against}

\item{distributions}{The distributions to test x against}
}
\value{
A dataframe with the distance and p value for each performed
ks test
}
\description{
This function gets the distance and p-value from a Kolmogorov-smirnov
test for selected distributions against a continuous input variable.
Possible distributions include "normal",
"lognormal", "gamma", "exponential", and "all".
}
