% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztils.R
\name{multiPDF_Z}
\alias{multiPDF_Z}
\title{multiPDF_Z}
\usage{
multiPDF_Z(x, seq_length, distributions, palette, var_name)
}
\arguments{
\item{x}{The variable to for which to plot PDFs}

\item{seq_length}{The number of points over which to fit x}

\item{distributions}{The distributions to fit x against}

\item{palette}{The color palette to use on the graph}

\item{var_name}{The variable name to use for x}
}
\value{
A plot showing the PDF of the selected variable against the selected distributions over the selected sequence length
}
\description{
This function is a version of MultiPDF_plot from my MultiFitR package that is customized to use
formatting that I like. However, this formatting uses several packages that I didn't want to include in the
base MultiFitR or MultiFitRgg packages.
}
