% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztils.R
\name{no_outliers}
\alias{no_outliers}
\title{No outliers}
\usage{
no_outliers(data, var)
}
\arguments{
\item{data}{The data to subset}

\item{var}{The variable to subset by}
}
\value{
A dataframe without entries containing outliers in
the selected variable.
}
\description{
This function returns a dataframe subsetted to not include observations that
are beyond the outliers of the specified variable.
}
\details{
Outliers are defined by the quantiles +- 1.5 times the interquartile range.
}
