% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztils.R
\name{glm_pseudoR2}
\alias{glm_pseudoR2}
\title{glm_pseudoR2}
\usage{
glm_pseudoR2(mod)
}
\arguments{
\item{mod}{The model for which to calculate the pseudoR^2}
}
\description{
A function for calculating the pseudo R^2 for a glm
}
