% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztils.R
\name{multiKS_cont}
\alias{multiKS_cont}
\title{Multiple Kolmogorov-Smirnov Tests for Continuous Variables}
\usage{
multiKS_cont(x, distributions)
}
\arguments{
\item{x}{The variable to perform KS tests against}

\item{distributions}{The distributions to test x against}
}
\value{
A dataframe with the distance and p value for each performed
KS test
}
\description{
This function gets the distance and p-value from a Kolmogorov-smirnov test for selected distributions
against a continuous, non-negative input variable. Possible distributions include "normal",
"lognormal", "gamma", "exponential", "cauchy", "t", "weibull", "logistic",
and "all".
}
