% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_extremes.R
\name{no_extremes}
\alias{no_extremes}
\title{No extremes}
\usage{
no_extremes(data, var)
}
\arguments{
\item{data}{The data to subset}

\item{var}{The variable to subset by.}
}
\value{
A dataframe without entries containing extremes in
the selected variable.
}
\description{
This function returns a dataframe subsetted to not include observations that
are beyond the extremes of the specified variable.
Extremes are defined by the quantiles +- 3 times the interquartile range.
}
\examples{
no_extremes(iris, Sepal.Length)

}
